# makes HAMT iteration 10x faster