The HAMT iterator completely redesigned

**before**
```
Target            Operation   Impl                         Params  Count            Score  ±     Error    Unit
HashSetBenchmark  Iterate     slang_persistent               1000     18        4,299.731  ±     1.18%   ops/s
HashSetBenchmark  Iterate     slang_persistent                100     18       45,524.153  ±     1.77%   ops/s
HashSetBenchmark  Iterate     slang_persistent                 10     18      680,479.197  ±     1.31%   ops/s
HashSetBenchmark  Iterate     scala_persistent               1000     18      129,055.550  ±     0.24%   ops/s
HashSetBenchmark  Iterate     scala_persistent                100     18    1,437,694.285  ±     0.83%   ops/s
HashSetBenchmark  Iterate     scala_persistent                 10     18   22,035,129.813  ±     1.27%   ops/s
HashSetBenchmark  Iterate     pcollections_persistent        1000     18       44,336.905  ±     1.08%   ops/s
HashSetBenchmark  Iterate     pcollections_persistent         100     18      472,947.893  ±     1.33%   ops/s
HashSetBenchmark  Iterate     pcollections_persistent          10     18    3,469,540.882  ±     0.70%   ops/s

Performance Ratios
================================================================================================================
  (Outliers removed: 30.00% low end, 10.00% high end)

Ratios slang / <alternative_impl>
Target            Operation   Ratio                                                    10        100       1000 
HashSetBenchmark  Iterate     slang_persistent/pcollections_persistent              0.20x      0.10x      0.10x 
HashSetBenchmark  Iterate     slang_persistent/scala_persistent                     0.03x      0.03x      0.03x 

Ratios <alternative_impl> / slang
Target            Operation                                             Ratio          10        100       1000 
HashSetBenchmark  Iterate            pcollections_persistent/slang_persistent       5.10x     10.39x     10.31x 
HashSetBenchmark  Iterate                   scala_persistent/slang_persistent      32.38x     31.58x     30.01x 
```

**after**

```
Target            Operation   Impl                         Params  Count            Score  ±     Error    Unit
HashSetBenchmark  Iterate     slang_persistent               1000     18       49,484.074  ±     0.26%   ops/s
HashSetBenchmark  Iterate     slang_persistent                100     18      522,937.431  ±     0.98%   ops/s
HashSetBenchmark  Iterate     slang_persistent                 10     18    4,728,144.061  ±     1.99%   ops/s
HashSetBenchmark  Iterate     scala_persistent               1000     18      130,241.725  ±     0.38%   ops/s
HashSetBenchmark  Iterate     scala_persistent                100     18    1,595,782.555  ±     1.03%   ops/s
HashSetBenchmark  Iterate     scala_persistent                 10     18   22,418,589.165  ±     1.87%   ops/s
HashSetBenchmark  Iterate     pcollections_persistent        1000     18       45,362.049  ±     1.55%   ops/s
HashSetBenchmark  Iterate     pcollections_persistent         100     18      471,520.387  ±     0.25%   ops/s
HashSetBenchmark  Iterate     pcollections_persistent          10     18    3,433,121.007  ±     3.10%   ops/s

Performance Ratios
================================================================================================================
  (Outliers removed: 30.00% low end, 10.00% high end)

Ratios slang / <alternative_impl>
Target            Operation   Ratio                                                    10        100       1000 
HashSetBenchmark  Iterate     slang_persistent/pcollections_persistent              1.38x      1.11x      1.09x 
HashSetBenchmark  Iterate     slang_persistent/scala_persistent                     0.21x      0.33x      0.38x 

Ratios <alternative_impl> / slang
Target            Operation                                             Ratio          10        100       1000 
HashSetBenchmark  Iterate            pcollections_persistent/slang_persistent       0.73x      0.90x      0.92x 
HashSetBenchmark  Iterate                   scala_persistent/slang_persistent       4.74x      3.05x      2.63x 
```